adding const and fixing all profile page route
