+++ capstone/lib/main.dart
@@ -19,9 +19,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return GetMaterialApp(
       getPages: AppPage.list,
-
-      initialRoute: AppRoute.splash_screen,
-
+      initialRoute: AppRoute.dashboard_screen,
       debugShowCheckedModeBanner: false,
       initialBinding: BindingsBuilder(() {
         Get.put<Auth>(Auth());
