+++ capstone/lib/route/app_route.dart
@@ -7,20 +7,25 @@ class AppRoute {
   static const String lupa_password = '/6';
   static const String dashboard_screen = '/7';
   static const String explore_screen = '/8';
-  static const String halamanukuranbarang_screen = '/9';
-  static const String hargabarang_screen = '/10';
-  static const String cariloker_screen = '/11';
-  static const String loker_tersedia_screen = '/12';
-  static const String loker_favorite_screen = '/13';
-  static const String durasi_penitipan_screen = '/14';
-  static const String category_barang_screen = '/15';
-  static const String proses_pemesanan_screen = '/16';
-  static const String promo_jasa_penitipan_screen = '/17';
-  static const String promo_pengiriman_screen = '/18';
-  static const String barang_berhasil_disimpan_screen = '/19';
-  static const String tracking_barang_screen = '/20';
-  static const String halaman_notifikasi_screen = '/21';
-  static const String estimasi_biaya_screen = '/22';
-  static const String ringkasan_pemesanan_page = '/23';
-  static const String metode_pembayaran_screen = '/24';
+  static const String profile_screen = '/9';
+  static const String biodata_screen = '/10';
+  static const String histori_pesanan_screen = '/11';
+  static const String pilih_bahasa_screen = '/12';
+  static const String faq_screen = '/13';
+  static const String halamanukuranbarang_screen = '/14';
+  static const String hargabarang_screen = '/15';
+  static const String cariloker_screen = '/16';
+  static const String loker_tersedia_screen = '/17';
+  static const String loker_favorite_screen = '/18';
+  static const String durasi_penitipan_screen = '/19';
+  static const String category_barang_screen = '/20';
+  static const String proses_pemesanan_screen = '/21';
+  static const String promo_jasa_penitipan_screen = '/22';
+  static const String promo_pengiriman_screen = '/23';
+  static const String barang_berhasil_disimpan_screen = '/24';
+  static const String tracking_barang_screen = '/25';
+  static const String halaman_notifikasi_screen = '/26';
+  static const String estimasi_biaya_screen = '/27';
+  static const String ringkasan_pemesanan_page = '/28';
+  static const String metode_pembayaran_screen = '/29';
 }
