+++ capstone/lib/view/profile/biodata_screen.dart
@@ -1,0 +1,457 @@
+import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
+
+import 'package:capstone/components/color_path.dart';
+
+class BiodataPage extends StatefulWidget {
+  const BiodataPage({super.key});
+
+  @override
+  State<BiodataPage> createState() => _BiodataPageState();
+}
+
+class _BiodataPageState extends State<BiodataPage> {
+  String? dropdownValue;
+
+  final ImagePicker _picker = ImagePicker();
+  List<XFile>? _imageFileList;
+  dynamic _pickImageError;
+
+  Future<void> _onImageButtonPressed(ImageSource source) async {
+    try {
+      final XFile? pickedFile = await _picker.pickImage(source: source);
+      setState(() {
+        _setImageFileListFromFile(pickedFile);
+      });
+    } catch (e) {
+      setState(() {
+        _pickImageError = e;
+      });
+    }
+  }
+
+  void _setImageFileListFromFile(XFile? file) {
+    if (file != null) {
+      setState(() {
+        _imageFileList = [file];
+      });
+    }
+  }
+
+  Widget bottomSheet() {
+    return Container(
+      height: 100.0,
+      width: 300.0,
+      margin: const EdgeInsets.symmetric(
+        horizontal: 20,
+        vertical: 20,
+      ),
+      child: Column(
+        children: [
+          const Text(
+            'Pilih Foto Profil Anda',
+            style: TextStyle(
+              fontSize: 20.0,
+            ),
+          ),
+          const SizedBox(
+            height: 20,
+          ),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              IconButton(
+                onPressed: () {
+                  _onImageButtonPressed(ImageSource.camera);
+                },
+                icon: const Icon(Icons.camera),
+              ),
+              const Text('Camera'),
+              const SizedBox(width: 50),
+              IconButton(
+                onPressed: () {
+                  _onImageButtonPressed(ImageSource.gallery);
+                },
+                icon: const Icon(Icons.image),
+              ),
+              const Text('Galeri'),
+            ],
+          )
+        ],
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: ColorPath.background2,
+      body: SafeArea(
+        child: Stack(
+          children: [
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                Expanded(
+                  child: ListView(
+                    padding: const EdgeInsets.all(20),
+                    children: [
+                      const Text(
+                        'Biodata',
+                        style: TextStyle(
+                          fontSize: 20,
+                          fontWeight: FontWeight.w600,
+                          color: ColorPath.background,
+                        ),
+                        textAlign: TextAlign.center,
+                      ),
+                      const SizedBox(height: 20),
+                      Stack(
+                        children: [
+                          const Row(
+                            children: [
+                              CircleAvatar(
+                                radius: 35.0,
+                                backgroundImage:
+                                    AssetImage('assets/images/macmiller.jpg'),
+                              ),
+                              SizedBox(width: 20),
+                              Text(
+                                'Alfian Adi Septianto',
+                                style: TextStyle(
+                                  fontFamily: 'Poppins',
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.w600,
+                                  color: ColorPath.background,
+                                ),
+                              ),
+                            ],
+                          ),
+                          Positioned(
+                            top: 40,
+                            left: 43,
+                            child: Container(
+                              width: 30,
+                              height: 30,
+                              decoration: BoxDecoration(
+                                  borderRadius: BorderRadius.circular(100),
+                                  color: ColorPath.background),
+                              child: IconButton(
+                                icon: const Icon(
+                                  Icons.camera_alt,
+                                  color: ColorPath.black,
+                                  size: 16,
+                                ),
+                                onPressed: () {
+                                  showModalBottomSheet(
+                                      context: context,
+                                      builder: (context) => bottomSheet());
+                                },
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                      const SizedBox(height: 20),
+                      Column(
+                        children: [
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'Nama Depan',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Container(
+                            width: 350,
+                            child: TextFormField(
+                              decoration: InputDecoration(
+                                hintText: 'Masukkan Nama Depan',
+                                hintStyle: const TextStyle(
+                                  fontSize: 13,
+                                  fontWeight: FontWeight.w500,
+                                  color: ColorPath.hinttext,
+                                ),
+                                border: OutlineInputBorder(
+                                  borderRadius: BorderRadius.circular(8.0),
+                                ),
+                                contentPadding: const EdgeInsets.symmetric(
+                                    horizontal: 16.0),
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'Nama Belakang',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Container(
+                            width: 350,
+                            child: TextFormField(
+                              decoration: InputDecoration(
+                                hintText: 'Masukkan Nama Belakang',
+                                hintStyle: const TextStyle(
+                                  fontSize: 13,
+                                  fontWeight: FontWeight.w500,
+                                  color: ColorPath.hinttext,
+                                ),
+                                border: OutlineInputBorder(
+                                  borderRadius: BorderRadius.circular(8.0),
+                                ),
+                                contentPadding: const EdgeInsets.symmetric(
+                                    horizontal: 16.0),
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'Jenis Kelamin',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Container(
+                            width: 350,
+                            height: 50,
+                            decoration: BoxDecoration(
+                              color: ColorPath.background2,
+                              border: Border.all(color: Colors.grey, width: 1),
+                              borderRadius: BorderRadius.circular(8),
+                            ),
+                            child: DropdownButtonFormField<String>(
+                              value: dropdownValue,
+                              items: <String>['Pria', 'Wanita']
+                                  .map((String value) {
+                                return DropdownMenuItem<String>(
+                                  value: value,
+                                  child: Text(value),
+                                );
+                              }).toList(),
+                              onChanged: (String? newValue) {
+                                setState(() {
+                                  dropdownValue = newValue;
+                                });
+                              },
+                              decoration: const InputDecoration(
+                                contentPadding: EdgeInsets.symmetric(
+                                    horizontal: 16, vertical: 10),
+                                enabledBorder: InputBorder.none,
+                                focusedBorder: InputBorder.none,
+                                hintText: 'Pilih Jenis Kelamin',
+                                hintStyle: TextStyle(
+                                  fontSize: 13,
+                                  fontWeight: FontWeight.w500,
+                                  color: ColorPath.hinttext,
+                                ),
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'Tanggal Lahir',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Row(
+                            children: [
+                              Container(
+                                width: 75,
+                                child: TextFormField(
+                                  decoration: InputDecoration(
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8.0),
+                                    ),
+                                    contentPadding: const EdgeInsets.symmetric(
+                                        horizontal: 16.0),
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(width: 30),
+                              Container(
+                                width: 75,
+                                child: TextFormField(
+                                  decoration: InputDecoration(
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8.0),
+                                    ),
+                                    contentPadding: const EdgeInsets.symmetric(
+                                        horizontal: 16.0),
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(width: 30),
+                              Container(
+                                width: 140,
+                                child: TextFormField(
+                                  decoration: InputDecoration(
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8.0),
+                                    ),
+                                    contentPadding: const EdgeInsets.symmetric(
+                                        horizontal: 16.0),
+                                  ),
+                                ),
+                              ),
+                            ],
+                          ),
+                          const SizedBox(height: 10),
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'No Telepon',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Row(
+                            children: [
+                              Container(
+                                width: 80,
+                                height: 50,
+                                decoration: BoxDecoration(
+                                  border:
+                                      Border.all(color: Colors.grey, width: 1),
+                                  color: const Color(0xFFEEF9F7),
+                                  borderRadius: BorderRadius.circular(8),
+                                  shape: BoxShape.rectangle,
+                                ),
+                                child: const Align(
+                                  alignment: AlignmentDirectional(0, 0),
+                                  child: Text(
+                                    '+62',
+                                    textAlign: TextAlign.center,
+                                    style: TextStyle(
+                                      fontWeight: FontWeight.w500,
+                                    ),
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(width: 30),
+                              Container(
+                                width: 240,
+                                child: TextFormField(
+                                  decoration: InputDecoration(
+                                    border: OutlineInputBorder(
+                                      borderRadius: BorderRadius.circular(8.0),
+                                    ),
+                                    contentPadding: const EdgeInsets.symmetric(
+                                        horizontal: 16.0),
+                                  ),
+                                ),
+                              ),
+                            ],
+                          ),
+                          const SizedBox(height: 10),
+                          const Align(
+                            alignment: Alignment.topLeft,
+                            child: Text(
+                              'Alamat',
+                              style: TextStyle(
+                                fontSize: 13,
+                                fontWeight: FontWeight.w600,
+                                color: ColorPath.background,
+                              ),
+                            ),
+                          ),
+                          const SizedBox(height: 10),
+                          Container(
+                            width: 350,
+                            child: TextFormField(
+                              decoration: InputDecoration(
+                                  hintText: 'Masukkan Lokasi',
+                                  hintStyle: const TextStyle(
+                                    fontSize: 13,
+                                    fontWeight: FontWeight.w500,
+                                    color: ColorPath.hinttext,
+                                  ),
+                                  border: OutlineInputBorder(
+                                    borderRadius: BorderRadius.circular(8.0),
+                                  ),
+                                  contentPadding: const EdgeInsets.symmetric(
+                                      horizontal: 16.0),
+                                  prefixIcon: const Icon(
+                                    Icons.place,
+                                    color: Color(0xFF02B3FF),
+                                  )),
+                            ),
+                          ),
+                          const SizedBox(height: 30),
+                          Container(
+                            width: 150,
+                            height: 50,
+                            decoration: BoxDecoration(
+                              border: Border.all(color: Colors.grey, width: 1),
+                              color: ColorPath.AppbarProfile,
+                              borderRadius: BorderRadius.circular(15),
+                              shape: BoxShape.rectangle,
+                            ),
+                            child: const Align(
+                              alignment: AlignmentDirectional(0, 0),
+                              child: Text(
+                                'Tambah',
+                                textAlign: TextAlign.center,
+                                style: TextStyle(
+                                  color: ColorPath.white,
+                                  fontWeight: FontWeight.w600,
+                                  fontSize: 16,
+                                ),
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                )
+              ],
+            ),
+            Positioned(
+              top: 10,
+              left: 16,
+              child: IconButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                icon: const Icon(
+                  Icons.arrow_back_ios,
+                  size: 24,
+                ),
+                color: Colors.black,
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
