+++ capstone/lib/view/profile/profile.dart
@@ -1,5 +1,5 @@
 export 'profile_screen.dart';
-export 'pilih_bahasa.dart';
-export 'histori_pesanan.dart';
-export 'favorit_loker.dart';
-export 'biodata.dart';
+export 'pilih_bahasa_screen.dart';
+export 'histori_pesanan_screen.dart';
+export 'loker_favorite_screen.dart';
+export 'biodata_screen.dart';
